created bids dataset module
