+++ arcana/dataset/__init__.py
@@ -4,3 +4,4 @@ from .spec import DatasetSpec, FieldSpec, BaseSpec
 from .base import BaseField, BaseDataset, BaseDatasetOrField
 from .match import DatasetMatch, FieldMatch, BaseMatch
 from .file_format import FileFormat, Converter, IdentityConverter
+from .bids import BidsDatasetMatch, BidsAssociatedDatasetMatch
